% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a}
\alias{a}
\title{Toy data}
\format{\preformatted{'data.frame':	3 obs. of  2 variables:
 $ x1: chr  "A" "B" "C"
 $ x2: int  1 2 3
}}
\usage{
a
}
\description{
Meaningless numbers and letters.
}
\keyword{datasets}

