% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artistas}
\alias{artistas}
\title{Nombres de musicos}
\format{An object of class \code{data.frame} with 4 rows and 2 columns.}
\usage{
artistas
}
\description{
Un conjunto de datos simple, consistiendo de nombres de músicos y los instrumentos
que tocan (o tocaban).
}
\keyword{datasets}

