% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artistas2}
\alias{artistas2}
\title{Nombres de musicos}
\format{An object of class \code{data.frame} with 6 rows and 3 columns.}
\usage{
artistas2
}
\description{
Un segundo conjunto de datos simple, consistiendo de nombres de músicos y
los instrumentos que tocan (o tocaban).
}
\keyword{datasets}

