% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artists2}
\alias{artists2}
\title{Names of musicians}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ first: chr  "George" "John" "Paul" "Ringo" ...
 $ last : chr  "Harrison" "Lennon" "McCartney" "Starr" ...
 $ plays: chr  "sitar" "guitar" "bass" "drums" ...
}}
\usage{
artists2
}
\description{
A second simple data set of musicians and the instruments they played.
}
\keyword{datasets}

