% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b}
\alias{b}
\title{Toy data}
\format{\preformatted{'data.frame':	3 obs. of  2 variables:
 $ x1: chr  "A" "B" "D"
 $ x2: logi  TRUE FALSE TRUE
}}
\usage{
b
}
\description{
Meaningless numbers and letters.
}
\keyword{datasets}

