% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canciones}
\alias{canciones}
\title{Nombres de canciones}
\format{An object of class \code{data.frame} with 4 rows and 2 columns.}
\usage{
canciones
}
\description{
Una lista simple de canciones pop y sus autores
}
\keyword{datasets}

