% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canciones2}
\alias{canciones2}
\title{Nombres de canciones}
\format{An object of class \code{data.frame} with 4 rows and 3 columns.}
\usage{
canciones2
}
\description{
Una segunda lista simple de canciones pop y sus autores
}
\keyword{datasets}

