% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{cases}
\alias{cases}
\title{TB cases in USA, Germany, and France}
\format{A dataset with the estimated number of TB cases in
France, Germany, and the United States for 2011, 2012, and
2013.}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
cases
}
\description{
A subset of data from the World Health Organization Global
Tuberculosis Report.
}
\keyword{datasets}

