% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casos}
\alias{casos}
\title{Casos de TB en EEUU, Alemania y Francia}
\format{Un conjunto de datos con el numero estimado de casos de TB en
Francia, Alemania y los EEUU para 2011, 2012 y 2013.}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
casos
}
\description{
Un subconjunto de datos del World Health Organization Global Tuberculosis
Report.
}
\keyword{datasets}

