% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mintb}
\alias{mintb}
\title{Datos mínimos TB}
\format{Un data frame con columnas
\describe{
\item{país}{}
\item{año}{}
\item{sexo}{}
\item{casos}{}
}}
\usage{
mintb
}
\description{
Un conjunto de datos mínimo del conjunto de datos \code{\link{tb}}. Los
datos muestran un numero imaginario de casos de TB agrupados por país, año y
sexo.
}
\keyword{datasets}

