% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nas}
\alias{nas}
\title{Toy data}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	4 obs. of  3 variables:
 $ a: num  NA 2 3 NA
 $ b: num  1 NA 3 NA
 $ c: num  1 2 NA NA
}}
\usage{
nas
}
\description{
A data set that has a tricky set of missing values to work with.
}
\keyword{datasets}

