% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutricion}
\alias{nutricion}
\title{Datos de nutrición}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8463 rows and 28 columns.}
\source{
\url{http://www.ars.usda.gov/Services/docs.htm?docid=8964}
}
\usage{
nutricion
}
\description{
Datos de nutrición de 8463 alimentos reunidos por el departamento de
agricultura de los EEUU (USDA). El conjunto de datos contiene la cantidad de
25 nutrientes y macronutrientes encontrados en 100 gramos de cada alimento,
así como el nombre, numero de calorías total y grupo alimenticio.
}
\details{
Los datos de nutrición fué compilado con base en conjuntos de datos en el
paquete \code{\link[usdanutrients]{usdanutrients}}.
}
\keyword{datasets}

