% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutrition}
\alias{nutrition}
\title{Nutrition data}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	8463 obs. of  28 variables:
 $ food         : chr  "Butter, salted" "Butter, whipped, with salt" "Butter oil, anhydrous" "Cheese, blue" ...
 $ calories     : num  717 717 876 353 371 334 300 376 403 387 ...
 $ protein      : num  0.85 0.85 0.28 21.4 23.24 ...
 $ carbohydrates: num  0.06 0.06 0 2.34 2.79 0.45 0.46 3.06 1.28 4.78 ...
 $ total_fat    : num  81.1 81.1 99.5 28.7 29.7 ...
 $ saturated_fat: num  51.4 50.5 61.9 18.7 18.8 ...
 $ caffiene     : num  0 0 0 0 0 0 0 NA 0 NA ...
 $ cholesterol  : num  215 219 256 75 94 100 72 93 105 103 ...
 $ fiber        : num  0 0 0 0 0 0 0 0 0 0 ...
 $ folic_acid   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ sodium       : num  643 659 2 1146 560 ...
 $ calcium      : num  24 24 4 528 674 184 388 673 721 643 ...
 $ iron         : num  0.02 0.16 0 0.31 0.43 0.5 0.33 0.64 0.68 0.21 ...
 $ magnesium    : num  2 2 0 23 24 20 20 22 28 21 ...
 $ manganese    : num  0 0.004 0 0.009 0.012 0.034 0.038 0.021 0.01 0.012 ...
 $ niacin       : num  0.042 0.042 0.003 1.016 0.118 ...
 $ phosphorus   : num  24 23 3 387 451 188 347 490 512 464 ...
 $ potassium    : num  24 26 5 256 136 152 187 93 98 95 ...
 $ riboflavin   : num  0.034 0.034 0.005 0.382 0.351 0.52 0.488 0.45 0.375 0.293 ...
 $ selenium     : num  1 1 0 14.5 14.5 14.5 14.5 14.5 13.9 14.5 ...
 $ thiamin      : num  0.005 0.005 0.001 0.029 0.014 0.07 0.028 0.031 0.027 0.046 ...
 $ vitamin_A    : num  2499 2499 3069 721 1080 ...
 $ vitamin_B6   : num  0.003 0.003 0.001 0.166 0.065 0.235 0.227 0.074 0.074 0.074 ...
 $ vitamin_B12  : num  0.17 0.13 0.01 1.22 1.26 1.65 1.3 0.27 0.83 0.83 ...
 $ vitamin_C    : num  0 0 0 0 0 0 0 0 0 0 ...
 $ vitamin_D    : num  60 60 73 21 22 20 18 NA 24 NA ...
 $ zinc         : num  0.09 0.05 0.01 2.66 2.6 2.38 2.38 2.94 3.11 2.79 ...
 $ group        : chr  "Dairy and Egg Products" "Dairy and Egg Products" "Dairy and Egg Products" "Dairy and Egg Products" ...
}}
\source{
\url{http://www.ars.usda.gov/Services/docs.htm?docid=8964}
}
\usage{
nutrition
}
\description{
Nutrition data on 8463 foods provided by the United States Department of Agriculture (USDA). The dataset contains the amount of 25 nutrients and macronutrients found in 100 grams of each food, as well as the name, total calories, and food group for each food.
}
\details{
The nutrition data was compiled from data sets in the \code{\link[usdanutrients]{usdanutrients}} package.
}
\keyword{datasets}

