% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{Vectorized addition with na.rm}
\usage{
plus(..., na.rm = TRUE)
}
\arguments{
\item{...}{a set of vectors. Each vector should be numeric, integer, or coercible to numeric.
na.rm logical. Should missing values (including NaN) be omitted from the calculations?}
}
\description{
\code{plus} adds together multiple vectors in element-wise fashion,
so \code{plus(a, b, c)} would be similar to \code{a + b +c}.
Unlike \code{+}, \code{plus} takes an \code{na.rm} argument to handle
NA behavior.
}
\examples{
a <- c(NA, 2, 3)
b <- c(1, NA, 3)
c <- c(1, 2, NA)
plus(a, b, c)
plus(a, b, c, na.rm = FALSE)
}

