% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poblacion}
\alias{poblacion}
\title{Datos de población}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 20 columns.}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
poblacion
}
\description{
Población de 100 países para 1995-2013 presentados en formato tabla.
}
\keyword{datasets}

