% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollution}
\alias{pollution}
\title{Pollution data}
\format{A data frame with variables:
\describe{
\item{city}{Name of city: New York, London, or Beijing}
\item{Size}{Size of air particulate measured. Fine suspended particles smaller than 10 microns in diameter (\code{large}) and 2.5 microns in diameter (\code{small}).}
\item{amount}{The mean annual concentration of particles in milligrams per meter cubed (ug/m3)}
}}
\source{
\url{http://www.who.int/phe/health_topics/outdoorair/databases/cities/en/}
}
\usage{
pollution
}
\description{
Pollution data from the WHO, 2014. This dataset contains
a subset of data from the Ambient Air Pollution Database,
WHO, May 2014.
}
\keyword{datasets}

