% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polucion}
\alias{polucion}
\title{Datos de polución}
\format{Un marco de datos con variables: 
\describe{
\item{ciudad}{Nombre de la ciudad: New York (Nueva York), London (Londres),
o Beijing}
\item{tamaño}{Tamaño del aire particulado medido. Finas partículas
suspendidas de tamaño menor a 10 micrones en diámetro, (\code{large}) y 2.5
micrones en diámetro (\code{small}).} 
\item{cantidad}{La cantidad anual promedio de concentración de partículas en
miligramos por metro cúbico (ug/m3)}
}}
\source{
\url{http://www.who.int/phe/health_topics/outdoorair/databases/cities/en/}
}
\usage{
polucion
}
\description{
Datos de polución del WHO, 2014. Este conjunto de datos contiene un
subconjunto de los datos del Ambient Air Pollution Database, WHO, Mayo 2014.
}
\keyword{datasets}

