% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population data}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	100 obs. of  20 variables:
 $ country: chr  "Afghanistan" "Algeria" "Angola" "Argentina" ...
 $ 1995   : int  17586073 29315463 12104952 34833168 7770806 119869585 10189075 5985658 7635362 1583453 ...
 $ 1996   : int  18415307 29845208 12451945 35264070 7852273 122400896 10156258 6176318 7806953 1620989 ...
 $ 1997   : int  19021226 30345466 12791388 35690778 7921745 124945315 10120556 6361301 7978521 1657349 ...
 $ 1998   : int  19496836 30820435 13137542 36109342 7984460 127478524 10080772 6546493 8150214 1692148 ...
 $ 1999   : int  19987071 31276295 13510616 36514558 8047936 129966823 10034775 6740491 8322408 1724924 ...
 $ 2000   : int  20595360 31719449 13924930 36903067 8117742 132383265 9981460 6949366 8495271 1755375 ...
 $ 2001   : int  21347782 32150198 14385283 37273361 8195427 134729503 9920173 7174911 8669066 1783349 ...
 $ 2002   : int  22202806 32572977 14886574 37627545 8279768 137006279 9852866 7414744 8843350 1808976 ...
 $ 2003   : int  23116142 33003442 15421075 37970411 8370169 139185986 9784111 7665681 9016787 1832602 ...
 $ 2004   : int  24018682 33461345 15976715 38308779 8465127 141235035 9719991 7922796 9187610 1854739 ...
 $ 2005   : int  24860855 33960903 16544376 38647854 8563398 143135180 9664714 8182362 9354709 1875805 ...
 $ 2006   : int  25631282 34507214 17122409 38988923 8665006 144868702 9620359 8443671 9517395 1895944 ...
 $ 2007   : int  26349243 35097043 17712824 39331357 8770122 146457067 9585368 8707490 9676456 1915187 ...
 $ 2008   : int  27032197 35725377 18314441 39676083 8877669 147969967 9555801 8973293 9834098 1933719 ...
 $ 2009   : int  27708187 36383302 18926650 40023641 8986266 149503100 9525785 9240783 9993406 1951715 ...
 $ 2010   : int  28397812 37062820 19549124 40374224 9094718 151125475 9491070 9509798 10156601 1969341 ...
 $ 2011   : int  29105480 37762962 20180490 40728738 9202432 152862431 9450391 9779795 10324445 1986701 ...
 $ 2012   : int  29824536 38481705 20820525 41086927 9308959 154695368 9405097 10050702 10496285 2003910 ...
 $ 2013   : int  30551674 39208194 21471618 41446246 9413420 156594962 9356678 10323474 10671200 2021144 ...
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
population
}
\description{
Populations of 100 countries for 1995-2013 laid out in table format
}
\keyword{datasets}

