% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{TB rates}
\format{A dataset with the variables
\describe{
\item{country}{}
\item{year}{}
\item{cases}{Number of new cases of TB reported}
\item{population}{}
\item{rate}{Number of new cases of TB reported per 10,000 people.}
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
rates
}
\description{
A subset of data from the World Health Organization Global
Tuberculosis Report, manipulated to show TB rates by country by year.
}
\keyword{datasets}

