% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{rawtb}
\alias{rawtb}
\title{Raw TB data}
\format{A dataset with the variables
\describe{
\item{country}{}
\item{year}{}
\item{sex}{}
\item{age}{One of \code{child} (0 - 14 years of age), \code{adult} (15 - 64 years of age), or \code{elderly} (over 64 years of age).}
\item{n}{Number of cases: in this layout, one.}
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
rawtb
}
\description{
A subset of the data in the \code{\link{tb}} data set. The data displays each individual TB case from Afghanistan, Algeria, Angola, Argentina, Azerbaijan, Belarus, Benin, Botswana, and Burundi in its own row.
}
\keyword{datasets}

