% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{songs}
\alias{songs}
\title{Names of songs}
\format{\preformatted{'data.frame':	4 obs. of  2 variables:
 $ song: chr  "Across the Universe" "Come Together" "Hello, Goodbye" "Peggy Sue"
 $ name: chr  "John" "John" "Paul" "Buddy"
}}
\usage{
songs
}
\description{
A simple list of pop songs and their authors
}
\keyword{datasets}

