% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{songs2}
\alias{songs2}
\title{Names of songs}
\format{\preformatted{'data.frame':	4 obs. of  3 variables:
 $ song : chr  "Across the Universe" "Come Together" "Hello, Goodbye" "Peggy Sue"
 $ first: chr  "John" "John" "Paul" "Buddy"
 $ last : chr  "Lennon" "Lennon" "McCartney" "Holly"
}}
\usage{
songs2
}
\description{
A second simple list of pop songs and their authors
}
\keyword{datasets}

