% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{storms}
\alias{storms}
\title{Hurricane data}
\format{A data frame with variables:
\describe{
\item{storm}{Name of hurricane: Alberto, Alex, Allison, Ana, Arlene, Arthur.}
\item{wind}{Maximum wind speed measured for each hurricane (in miles per hour).}
\item{pressure}{Air pressure in the storm's center (in millibars).}
\item{date}{Date maximum wind speed was observed.}
}}
\source{
\url{http://www.nhc.noaa.gov/}
}
\usage{
storms
}
\description{
Wind speed data for six hurricanes, collected from
National Hurricane Center's archive of Tropical Cyclone
Reports.
}
\keyword{datasets}

