% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasas}
\alias{tasas}
\title{Tasas de TB}
\format{Un conjunto de datos con variables
\describe{
\item{país}{}
\item{año}{}
\item{casos}{Numero de nuevos casos de TB reportados}
\item{populación}{}
\item{tasa}{Numero de nuevos casos de TB reportados por cada 10,000
personas.}
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
tasas
}
\description{
Un subconjunto de datos del informe sobre Tuberculosis de la Organización
Mundial de Salud, manipulado para mostrar las tasas de TB por país por año.
}
\keyword{datasets}

