% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb}
\alias{tb}
\title{TB data}
\format{A dataset with the variables
\describe{
\item{country}{}
\item{year}{}
\item{sex}{}
\item{child}{Number of new cases reported among people 0 - 14 years of age.}
\item{adult}{Number of new cases reported among people 15 - 64 years of age.}
\item{elderly}{Number of new cases reported among people over 64 years of age.}
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
tb
}
\description{
A subset of data from the World Health Organization Global
Tuberculosis Report.
}
\keyword{datasets}

