% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbcrudo}
\alias{tbcrudo}
\title{Datos crudos TB}
\format{Un conjunto de datos con variables
\describe{
\item{país}{}
\item{año}{}
\item{sexo}{}
\item{edad}{Uno de \code{niño} (0 - 14 años de edad), \code{adulto} (15 - 64
años de edad), o \code{mayor} (mas de 64 años de edad).}
\item{n}{Numero de casos: in en esta vista, uno.}
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
tbcrudo
}
\description{
Un subconjunto de los datos en \code{\link{tb}}. Los datos muestran cada
caso individual de Afganistán, Argelia, Angola, Argentina, Azerbaiyán,
Bielorrusia, Benin, Botsuana y Burundi en su propia fila.
}
\keyword{datasets}

