% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidypop}
\alias{tidypop}
\title{Population data}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1900 obs. of  3 variables:
 $ country   : chr  "Afghanistan" "Algeria" "Angola" "Argentina" ...
 $ year      : int  1995 1995 1995 1995 1995 1995 1995 1995 1995 1995 ...
 $ population: int  17586073 29315463 12104952 34833168 7770806 119869585 10189075 5985658 7635362 1583453 ...
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
tidypop
}
\description{
Populations of 100 countries for 1995-2013 laid out in tidy format
}
\keyword{datasets}

