% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tormentas}
\alias{tormentas}
\title{Datos de huracanes}
\format{Un marco de datos con variables: 
\describe{
\item{tormenta}{Nombre del huracán: Alberto, Alex, Allison, Ana, Arlene,
Arthur.}
\item{viento}{Velocidad máxima del viento medido para cada huracán (en
millas por hora)}
\item{presión}{Presión del aire en el centro de la tormenta (en milibares).}
\item{fecha}{Fecha en la cual se observó la velocidad de viento máxima.}
}}
\source{
\url{http://www.nhc.noaa.gov/}
}
\usage{
tormentas
}
\description{
Velocidades de viento para seis huracanes, reunidos del archivo de Centro
Nacional de Huracanes en sus informes sobre Ciclones Tropicales.
}
\keyword{datasets}

