% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyb}
\alias{toyb}
\title{Toy TB data}
\format{A data frame with the columns
\describe{
\item{country}{}
\item{year}{}
\item{sex}{}
\item{cases}{}
}}
\usage{
toyb
}
\description{
A toy data set based on the \code{\link{tb}} data set. The data displays an imaginary number of TB cases grouped by country, year, and sex.
}
\keyword{datasets}

