% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{Toy data}
\format{\preformatted{'data.frame':	3 obs. of  2 variables:
 $ x1: Factor w/ 3 levels "A","B","C": 1 2 3
 $ x2: int  1 2 3
}}
\usage{
y
}
\description{
Meaningless numbers and letters.
}
\keyword{datasets}

