% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z}
\alias{z}
\title{Toy data}
\format{\preformatted{'data.frame':	3 obs. of  2 variables:
 $ x1: Factor w/ 3 levels "B","C","D": 1 2 3
 $ x2: int  2 3 4
}}
\usage{
z
}
\description{
Meaningless numbers and letters.
}
\keyword{datasets}

