% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{artists}
\alias{artists}
\title{Names of musicians}
\format{\preformatted{'data.frame':	4 obs. of  2 variables:
 $ name : Factor w/ 4 levels "George","John",..: 1 2 3 4
 $ plays: Factor w/ 4 levels "bass","drums",..: 4 3 1 2
}}
\usage{
artists
}
\description{
A simple data set of musicians and the instruments they played.
}
\keyword{datasets}

