% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artists}
\alias{artists}
\title{Names of musicians}
\format{\preformatted{'data.frame':	4 obs. of  2 variables:
 $ name : chr  "George" "John" "Paul" "Ringo"
 $ plays: chr  "sitar" "guitar" "bass" "drums"
}}
\usage{
artists
}
\description{
A simple data set of musicians and the instruments they played.
}
\keyword{datasets}

