% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artists2}
\alias{artists2}
\title{Names of musicians}
\format{An object of class \code{data.frame} with 6 rows and 3 columns.}
\usage{
artists2
}
\description{
A second simple data set of musicians and the instruments they played.
}
\keyword{datasets}

