% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{artists2}
\alias{artists2}
\title{Names of musicians}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ first: Factor w/ 4 levels "George","John",..: 1 2 3 4 3 2
 $ last : Factor w/ 6 levels "Coltrane","Harrison",..: 2 3 4 6 5 1
 $ plays: Factor w/ 5 levels "bass","drums",..: 5 3 1 2 3 4
}}
\usage{
artists2
}
\description{
A second simple data set of musicians and the instruments they played.
}
\keyword{datasets}

