% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nas}
\alias{nas}
\title{Datos mínimos}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.}
\usage{
nas
}
\description{
Un conjunto de datos una serie de datos faltantes mañosa.
}
\keyword{datasets}

