% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutrition}
\alias{nutrition}
\title{Nutrition data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8463 rows and 28 columns.}
\source{
\url{http://www.ars.usda.gov/Services/docs.htm?docid=8964}
}
\usage{
nutrition
}
\description{
Nutrition data on 8463 foods provided by the United States Department of Agriculture (USDA). The dataset contains the amount of 25 nutrients and macronutrients found in 100 grams of each food, as well as the name, total calories, and food group for each food.
}
\details{
The nutrition data was compiled from data sets in the \code{\link[usdanutrients]{usdanutrients}} package.
}
\keyword{datasets}

