% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pollution}
\alias{pollution}
\title{Pollution data}
\format{A data frame with variables:
\describe{
\item{city}{Name of city: New Yourk, London, or Beijing}
\item{type}{Type of air particulate measured. Fine suspended particles smaller than 10 microns in diameter (\code{PM_10}) and 2.5 microns in diameter (\code{PM_2.5}).}
\item{value}{The mean annual concentration of particles in milligrams per meter cubed (ug/m3)}
}}
\source{
\url{http://www.who.int/phe/health_topics/outdoorair/databases/cities/en/}
}
\usage{
pollution
}
\description{
Pollution data from the WHO, 2014. This dataset contains
a subset of data from the Ambient Air Pollution Database,
WHO, May 2014.
}
\keyword{datasets}

