% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 20 columns.}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
population
}
\description{
Populations of 100 countries for 1995-2013 laid out in table format
}
\keyword{datasets}

