% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{songs2}
\alias{songs2}
\title{Names of songs}
\format{An object of class \code{data.frame} with 4 rows and 3 columns.}
\usage{
songs2
}
\description{
A second simple list of pop songs and their authors
}
\keyword{datasets}

