% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{songs2}
\alias{songs2}
\title{Names of songs}
\format{\preformatted{'data.frame':	4 obs. of  3 variables:
 $ song : Factor w/ 4 levels "Across the Universe",..: 1 2 3 4
 $ first: Factor w/ 3 levels "Buddy","John",..: 2 2 3 1
 $ last : Factor w/ 3 levels "Holly","Lennon",..: 2 2 3 1
}}
\usage{
songs2
}
\description{
A second simple list of pop songs and their authors
}
\keyword{datasets}

