% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb}
\alias{tb}
\title{Datos TB}
\format{Un conjunto de datos con variables
\describe{
\item{país}{}
\item{año}{}
\item{sexo}{}
\item{niño}{Numero de casos nuevos reportados entre personas de 0 - 14 años de edad.}
\item{adulto}{Numero de casos nuevos reportados entre personas de 15 - 64 años de edad.}
\item{mayor}{Numero de casos nuevos reportados entre personas de más de 64 años de edad.}
}}
\source{
\url{http://www.who.int/tb/country/data/download/en/}
}
\usage{
tb
}
\description{
Un subconjunto de datos del World Health Organization Global Tuberculosis
Report.
}
\keyword{datasets}

